add README

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public int calculate(double weight, double height) {
+    public int calculate(int weight, double height) {
         double bmi = weight / (height * height);
         return (int) bmi;
     }
